requirements.txt
asyncio
aiohttp
psutil
discord.py
DateTime
numpy
gTTS
Pillow
jishaku